Better tests

+++ pkg/entry.sh
@@ -1,30 +1,34 @@
 #!/bin/bash
 
 if [ "$TEST_MODE" == "true" ]; then
-    echo "Running in test mode"
-    source "$(dirname $0)"/logger.sh # Use the logger script
-    source "$(dirname $0)"/utils.sh
-    source "$(dirname $0)"/env.sh 
+    echo "--- Running entry.sh in TEST mode ---"
+    source ../pkg/logger.sh # Use the logger script
+    source ../pkg/utils.sh
+    source ../pkg/env.sh
 else
     source /app/logger.sh # Use the logger script
     source /app/utils.sh
     source app/env.sh
 fi
 
-# Echo the CRON schedule for logging/debugging
-logThis "Installing CRON schedule: $CRON_SCHEDULE in TZ: $TZ" "DEBUG" "init"
+if [ "$TEST_MODE" == "false" ]; then
+    echo "Skipping CRON schedule installation in test mode"
 
-# Dump the current cron jobs to a temporary file
-crontab -l >tempcron
+    # Echo the CRON schedule for logging/debugging
+    logThis "Installing CRON schedule: $CRON_SCHEDULE in TZ: $TZ" "DEBUG" "init"
 
-# Remove the existing cron job for your backup script from the file
-sed -i '/\/app\/backup.sh/d' tempcron
+    # Dump the current cron jobs to a temporary file
+    crontab -l >tempcron
 
-# Add the new cron job to the file
-echo "$CRON_SCHEDULE bash /app/backup.sh" >>tempcron
+    # Remove the existing cron job for your backup script from the file
+    sed -i '/\/app\/backup.sh/d' tempcron
 
-# Install the new cron jobs and remove the tempcron file
-crontab tempcron && rm tempcron
+    # Add the new cron job to the file
+    echo "$CRON_SCHEDULE bash /app/backup.sh" >>tempcron
+
+    # Install the new cron jobs and remove the tempcron file
+    crontab tempcron && rm tempcron
+fi
 
 # Verify the source and destination locations
 verify_source_location $SOURCE_LOCATION
@@ -36,10 +40,12 @@ if [ "$BACKUP_ON_START" = "true" ]; then
 fi
 
 if [ "$EXIT_AFTER_INIT" = "true" ]; then
-    echo "Exiting since EXIT_AFTER_INITis true" "INFO" "init"
+    echo "Exiting since EXIT_AFTER_INIT is true" "INFO" "init"
     exit 0
 fi
 
 logThis "Initialization complete. Awaiting CRON schedule: $CRON_SCHEDULE" "INFO" "init"
 
-/usr/sbin/crond -f -l 8 # Start cron and keep container running
+if [ "$TEST_MODE" == "false" ]; then
+    /usr/sbin/crond -f -l 8 # Start cron and keep container running
+fi
